EICNET-2802: Refactor the Profile rest resource.
