+++ config/sync/rest.resource.eic_webservices_profile.yml
@@ -1,16 +1,19 @@
-uuid: 3a966ef8-a693-4c44-b943-8f1feffd14ad
+uuid: aa63a2e3-4ec6-4044-bfd5-ff96d32cc1b0
 langcode: en
 status: true
 dependencies:
   module:
     - eic_webservices
     - hal
-id: eic_webservices_profile_update
-plugin_id: eic_webservices_profile_update
+    - profile
+id: eic_webservices_profile
+plugin_id: eic_webservices_profile
 granularity: resource
 configuration:
   methods:
+    - GET
     - POST
+    - PATCH
   formats:
     - hal_json
   authentication:
