+++ lib/modules/eic_webservices/config/install/rest.resource.eic_webservices_profile.yml
@@ -1,0 +1,19 @@
+langcode: en
+status: true
+dependencies:
+  module:
+    - eic_webservices
+    - hal
+    - user
+id: eic_webservices_profile
+plugin_id: eic_webservices_profile
+granularity: resource
+configuration:
+  methods:
+    - GET
+    - PATCH
+    - POST
+  formats:
+    - hal_json
+  authentication:
+    - eic_webservices_token_auth
